Make sure avro ints and floats are always written properly; fixes #522 (#532)

+++ include/rfl/avro/schema/Type.hpp
@@ -6,7 +6,6 @@
 #include <string>
 
 #include "../../Literal.hpp"
-//#include "../../Object.hpp"
 #include "../../Ref.hpp"
 #include "../../Rename.hpp"
 #include "../../Variant.hpp"
@@ -36,7 +35,7 @@ struct RFL_API Type {
   };
 
   struct Double {
-    Literal<"float"> type{};
+    Literal<"double"> type{};
   };
 
   struct Bytes {
